Fixed bug with negative centralities
In massive graphs, the centrality can be so large that an overflow error occurs. I changed the variable type to longs. Now this will accompany star graphs of over 2 billion nodes (star graphs would be the worst case scenario).

+++ GDFBetweenness.java
@@ -42,7 +42,7 @@ public class GDFBetweenness {
         long name;
         AttribNode parent;
         double distance;
-        int centrality;
+        long centrality;
 
 //        double x;  //EDITABLE add variables here if you need attributes in your calculations
 //        double y;
